Turned into electron app (i hate chrome)
