+++ src/stylesheets/style.css
@@ -1,4 +1,7 @@
-
+@font-face {
+    font-family: "Roboto";
+    src: url("../../assets/Roboto-Regular.ttf");
+}
 
 body {
     margin: 0;
@@ -36,7 +39,7 @@ body {
     background-color: #fff;
     box-shadow: 0 0 7px rgba(0, 0, 0, 0.15);
     resize: none;
-    font-family: 'Roboto', sans-serif;
+    font-family: 'Roboto';
     font-size: 16px;
     transition: transform 0.3s, opacity 0.3s, outline 0.3s;
 
@@ -58,7 +61,7 @@ body {
     padding: 16px;
     z-index: 1;
     font-size: 16px;
-    font-family: 'Roboto', sans-serif;
+    font-family: 'Roboto';
     color: #000;
     overflow-wrap: anywhere; /* Generate new line within words */
     overflow-wrap: break-word; /* Fallback for browsers that don't support 'anywhere' */
@@ -81,7 +84,7 @@ body {
     left: 0;
     padding: 5px;
     text-align: left;
-    font-family: 'Roboto', sans-serif;
+    font-family: 'Roboto';
     color: rgba(46, 46, 46, 0.15);
 }
 
@@ -129,7 +132,7 @@ body {
     padding-left: 24px;
     padding-right: 24px;
     color: #FFF;
-    font-family: 'Roboto', sans-serif;
+    font-family: 'Roboto';
     font-size: 25px;
 }
 
@@ -209,14 +212,17 @@ header {
 }
 
 #sidebar-text {
-    font-family: 'Roboto', sans-serif;
+    font-family: 'Roboto';
     color: #2e2e2e;
+    overflow-wrap: break-word;
+    word-wrap: break-word;
+    word-break: break-word;
 }
   
 .sidebar a {
     text-decoration: none;
     color: #000;
-    font-family: 'Roboto', sans-serif;
+    font-family: 'Roboto';
     font-size: 16px;
 
 }
@@ -229,7 +235,7 @@ header {
   
 .sidebar-header p {
     /*margin: 10px;*/
-    font-family: 'Roboto', sans-serif;
+    font-family: 'Roboto';
     font-weight: bold;
     font-size: 25px;
 }
